# Generates typeIDs at compilation time