--- a/builder.go
+++ b/builder.go
@@ -328,6 +328,7 @@ func (b *hostModuleBuilder) Compile(ctx context.Context) (CompiledModule, error)
 		return nil, err
 	}
 
+	// typeIDs are static and compile-time known.
 	typeIDs, err := b.r.store.GetFunctionTypeIDs(module.TypeSection)
 	if err != nil {
 		return nil, err