Previously we generated the typeIDs for function types unnecessarily at each instantiation time. 
TypeIDs are unique per Runtime/Store, and we don't need to create slice []FunctionTypeID for each instance.
When generating a typeID, we need to take a lock on store in order to keep uniqueness across
instances and this caused the lock contention on the instantiation heavy use-cases.

tldr is that this improves the concurrent instantiation performance!

```
$ benchstat main.txt new.txt
name                                    old time/op    new time/op    delta
Initialization/interpreter-10             29.9µs ± 3%    29.6µs ± 1%     ~     (p=0.151 n=5+5)
Initialization/interpreter-multiple-10    17.1µs ± 7%    13.2µs ± 9%  -23.13%  (p=0.008 n=5+5)
Initialization/compiler-10                35.1µs ± 3%    27.2µs ±47%     ~     (p=0.905 n=4+5)
Initialization/compiler-multiple-10       14.8µs ±15%    13.0µs ± 6%  -12.50%  (p=0.008 n=5+5)
Compilation/with_extern_cache-10           148µs ± 1%     152µs ± 3%     ~     (p=0.056 n=5+5)
Compilation/without_extern_cache-10       2.76ms ± 1%    2.74ms ± 2%     ~     (p=0.151 n=5+5)

name                                    old alloc/op   new alloc/op   delta
Initialization/interpreter-10              137kB ± 0%     137kB ± 0%   -0.04%  (p=0.000 n=5+4)
Initialization/interpreter-multiple-10     137kB ± 0%     137kB ± 0%     ~     (p=0.159 n=4+5)
Initialization/compiler-10                 141kB ± 0%     141kB ± 0%   -0.03%  (p=0.029 n=4+4)
Initialization/compiler-multiple-10        142kB ± 0%     142kB ± 0%   -0.10%  (p=0.008 n=5+5)
Compilation/with_extern_cache-10          55.4kB ± 0%    55.5kB ± 0%   +0.22%  (p=0.008 n=5+5)
Compilation/without_extern_cache-10       1.12MB ± 0%    1.12MB ± 0%     ~     (p=1.000 n=5+5)

name                                    old allocs/op  new allocs/op  delta
Initialization/interpreter-10               52.0 ± 0%      51.0 ± 0%   -1.92%  (p=0.008 n=5+5)
Initialization/interpreter-multiple-10      58.0 ± 0%      57.0 ± 0%   -1.72%  (p=0.029 n=4+4)
Initialization/compiler-10                  42.0 ± 0%      41.0 ± 0%   -2.38%  (p=0.008 n=5+5)
Initialization/compiler-multiple-10         48.0 ± 0%      47.0 ± 0%   -2.08%  (p=0.008 n=5+5)
Compilation/with_extern_cache-10           1.10k ± 0%     1.10k ± 0%   +0.18%  (p=0.008 n=5+5)
Compilation/without_extern_cache-10        11.4k ± 0%     11.4k ± 0%     ~     (p=0.079 n=5+5)
```